# docs/swagger.yaml
basePath: /api
definitions:
...
  product.CreateRequest:
    properties:
      alias:
        type: string
      product_info:
        $ref: '#/definitions/product.Info'
    type: object
  product.GetResponse:
    properties:
      error:
        type: string
      products:
        items:
          $ref: '#/definitions/storage.ProductInfo'
        type: array
      status:
        type: string
    type: object
  product.Info:
    properties:
      description:
        type: string
      image_id:
        type: integer
      name:
        type: string
      price:
        type: integer
    type: object
  response.Response:
    properties:
      error:
        type: string
      status:
        type: string
    type: object
...
  storage.ProductInfo:
    properties:
      description:
        type: string
      id:
        type: integer
      image_id:
        type: integer
      name:
        type: string
      price:
        type: integer
      website_id:
        type: integer
    type: object
  website.AliasesResponse:
    properties:
      aliases:
        items:
          type: string
        type: array
      error:
        type: string
      status:
        type: string
    type: object
  website.CreateRequest:
    properties:
      alias:
        type: string
    type: object
...
host: localhost:8082
info:
  contact: {}
  title: ForeignKey
  version: "1.0"
paths:
...
  /product/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: alias new website
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/product.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Create product
      tags:
      - product
  /product/get-by-alias/{alias}:
    get:
      description: Возвращает все товары сайта по алиасу
      parameters:
      - description: website alias
        in: path
        name: alias
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product.GetResponse'
      summary: GetByAlias
      tags:
      - product
  /website/aliases:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/website.AliasesResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all users aliases
      tags:
      - website
  /website/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: alias new website
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/website.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Create website
      tags:
      - website
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"